### 2020-12-08 #2896 @whedon
[PRE REVIEW]: lidar: A Python package for terrain and hydrological analysis using digital elevation models
**Submitting author:** @giswqs (<a href="http://orcid.org/0000-0001-5437-4073">Qiusheng Wu</a>)
**Repository:** <a href="https://github.com/giswqs/lidar" target ="_blank">https://github.com/giswqs/lidar</a>
**Version:** 0.5.1
