### 2020-12-08 #2896 @Kevin-Mattheus-Moerman
@giswqs since this submission is on the small side (in therms of number of lines of code). I have flagged this submission for an additional scope check by the editorial board. In particular our board will now discuss if it meets the [substantial scholarly effort](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort) criterion for review by JOSS. We should get back to you sometime next week.
