+++ lidar/toolbox/scripts/3_Flow_Path.py
@@ -1,0 +1,283 @@
+import arcpy
+import os
+import sys
+import time
+import string
+from arcpy import env
+import collections
+
+
+def FlowPath(in_dem, in_sink, rain_intensity, out_flowpath):
+    arcpy.CheckOutExtension("Spatial")
+    workspace = os.path.split(out_flowpath)[0]
+    arcpy.env.workspace = workspace
+    arcpy.env.overwriteOutput = True
+    dem = arcpy.Raster(in_dem)
+    cell_size = dem.meanCellWidth
+
+    if arcpy.Exists(in_dem) == False:
+        arcpy.AddMessage("The input raster does not exist")
+        quit()
+
+    if (os.path.splitext(out_flowpath)[1].lower() == ".shp"):
+        FieldOID = "FID"
+        FlowDir = os.path.join(workspace,"FlowDir.tif")
+        SinkCentroid = os.path.join(workspace,"SinkCentroid.shp")
+        CostPath = os.path.join(workspace,"CostPath.tif")
+        PathThin = os.path.join(workspace,"PathThin.tif")
+        PathLine = os.path.join(workspace,"PathLine.shp")
+        PathLineErase = os.path.join(workspace,"PathLineErase.shp")
+        Path = os.path.join(workspace,"FlowPath_Raw.shp")
+        # Path = out_flowpath
+        LineFlip = os.path.join(workspace,"LineFlip.shp")
+        LineNoFlip = os.path.join(workspace,"LineNoFlip.shp")
+        FlowFrom = os.path.join(workspace,"FlowFrom.shp")
+        FlowTo = os.path.join(workspace,"FlowTo.shp")
+        PathLineEraseSingle = os.path.join(workspace,"PathLineEraseSingle.shp")
+        LineStart = os.path.join(workspace,"LineStart.shp")
+        LineEnd = os.path.join(workspace, "LineEnd.shp")
+        LineStartElev = os.path.join(workspace,"LineStartElev.shp")
+        LineEndElev = os.path.join(workspace, "LineEndElev.shp")
+        PathBuffer = os.path.join(workspace,"PathBuffer.shp")
+        PathBufferSingle = os.path.join(workspace,"PathBufferSingle.shp")
+        FlowFromJoin = os.path.join(workspace, "FlowFromJoin.shp")
+        FlowToJoin = os.path.join(workspace, "FlowToJoin.shp")
+        FlowFromJoinBuffer = os.path.join(workspace, "FlowFromJoinBuffer.shp")
+        FlowToJoinBuffer = os.path.join(workspace, "FlowToJoinBuffer.shp")
+
+    else:
+        FieldOID = "OBJECTID"
+        FlowDir = os.path.join(workspace,"FlowDir")
+        SinkCentroid = os.path.join(workspace,"SinkCentroid")
+        CostPath = os.path.join(workspace,"CostPath")
+        PathThin = os.path.join(workspace,"PathThin")
+        PathLine = os.path.join(workspace,"PathLine")
+        PathLineErase = os.path.join(workspace,"PathLineErase")
+        Path = os.path.join(workspace,"FlowPath")
+        LineFlip = os.path.join(workspace,"LineFlip")
+        LineNoFlip = os.path.join(workspace,"LineNoFlip")
+        FlowFrom = os.path.join(workspace,"FlowFrom")
+        FlowTo = os.path.join(workspace,"FlowTo")
+        LineStart = os.path.join(workspace, "LineStart.shp")
+        LineEnd = os.path.join(workspace, "LineEnd.shp")
+    ### Delineate flow direction
+    flow_dir = arcpy.sa.FlowDirection(in_dem)
[…]
