### 2020-12-12 #2896 @giswqs
@Kevin-Mattheus-Moerman Just let you know that I have added more functionality to the package and expanded the source code by a few more hundreds of lines. You can rerun the statistics if needed. Thanks.
