### 2020-12-17 #2896 @kbarnhart
@kthyng I can handle this starting in early Jan.

@giswqs thanks for the submission. It still has the query-scope label, which means we are still determining whether the submission fits the scope of JOSS. I may follow up with some questions/clarifications on this point in early January. Please let me know if you have questions.
