### 2020-12-29 #2896 @kyleniemeyer
@kbarnhart @giswqs the editorial board voted to consider this one in scope, so please proceed! I'm removing the `scope-query` label.
