### 2021-01-04 #2896 @kbarnhart
@giswqs I'm now back in the office and will start actively handling this submission.

A couple of clarifying questions.

1. Is there a place where API-level documentation is provided? I've looked through the readthedocs page and can't seem to find this. I would expect that all public facing functions/classes would have detailed documentation describing the format of all inputs and outputs.
2. Could you comment on the similarities/difference between the core package and the arcpy toolbox. E.g., is
lidar/toolbox/scripts/4_Slicing.py intended as the arcpy equivalent of lidar/slicing.py?
3. It appears that the tests in lidar/tests are empty. Could you comment on how the contents of the package is tested?

A final comment: There are a number of dependencies (matplotlib, scipy, numpy, richDEM, skimage) that are used and not cited in the submitted paper.
