### 2021-01-05 e6ea2be
Updated JOSS paper
