+++ paper/paper.bib
@@ -129,3 +129,113 @@
   issn      = "0885-6087",
   doi       = "10.1002/hyp.10648"
 }
+
+@ARTICLE{Harris2020,
+  title     = "{Array programming with NumPy}",
+  author    = "Harris, Charles R and Millman, K Jarrod and van der Walt,
+               St{\'e}fan J and Gommers, Ralf and Virtanen, Pauli and
+               Cournapeau, David and Wieser, Eric and Taylor, Julian and Berg,
+               Sebastian and Smith, Nathaniel J and Kern, Robert and Picus,
+               Matti and Hoyer, Stephan and van Kerkwijk, Marten H and Brett,
+               Matthew and Haldane, Allan and Del R{\'\i}o, Jaime Fern{\'a}ndez
+               and Wiebe, Mark and Peterson, Pearu and G{\'e}rard-Marchant,
+               Pierre and Sheppard, Kevin and Reddy, Tyler and Weckesser,
+               Warren and Abbasi, Hameer and Gohlke, Christoph and Oliphant,
+               Travis E",
+  journal   = "Nature",
+  publisher = "nature.com",
+  volume    =  585,
+  number    =  7825,
+  pages     = "357--362",
+  month     =  sep,
+  year      =  2020,
+  url       = "http://dx.doi.org/10.1038/s41586-020-2649-2",
+  language  = "en",
+  issn      = "0028-0836, 1476-4687",
+  pmid      = "32939066",
+  doi       = "10.1038/s41586-020-2649-2"
+}
+
+@ARTICLE{Virtanen2020,
+  title     = "{SciPy 1.0: fundamental algorithms for scientific computing in
+               Python}",
+  author    = "Virtanen, Pauli and Gommers, Ralf and Oliphant, Travis E and
+               Haberland, Matt and Reddy, Tyler and Cournapeau, David and
+               Burovski, Evgeni and Peterson, Pearu and Weckesser, Warren and
+               Bright, Jonathan and van der Walt, St{\'e}fan J and Brett,
+               Matthew and Wilson, Joshua and Millman, K Jarrod and Mayorov,
+               Nikolay and Nelson, Andrew R J and Jones, Eric and Kern, Robert
+               and Larson, Eric and Carey, C J and Polat, {\.I}lhan and Feng,
+               Yu and Moore, Eric W and VanderPlas, Jake and Laxalde, Denis and
+               Perktold, Josef and Cimrman, Robert and Henriksen, Ian and
+               Quintero, E A and Harris, Charles R and Archibald, Anne M and
+               Ribeiro, Ant{\^o}nio H and Pedregosa, Fabian and van Mulbregt,
+               Paul and {SciPy 1.0 Contributors}",
+  journal   = "Nature methods",
+  publisher = "nature.com",
+  volume    =  17,
+  number    =  3,
+  pages     = "261--272",
+  month     =  mar,
+  year      =  2020,
+  url       = "http://dx.doi.org/10.1038/s41592-019-0686-2",
+  language  = "en",
+  issn      = "1548-7091, 1548-7105",
+  pmid      = "32015543",
+  doi       = "10.1038/s41592-019-0686-2",
+  pmc       = "PMC7056644"
+}
+
+@ARTICLE{Van_der_Walt2014,
+  title    = "{scikit-image: image processing in Python}",
+  author   = "van der Walt, St{\'e}fan and Sch{\"o}nberger, Johannes L and
+              Nunez-Iglesias, Juan and Boulogne, Fran{\c c}ois and Warner,
+              Joshua D and Yager, Neil and Gouillart, Emmanuelle and Yu, Tony
+              and {scikit-image contributors}",
+  journal  = "PeerJ",
+  volume   =  2,
+  pages    = "e453",
[…]
