+++ docs/lidar.rst
@@ -1,0 +1,47 @@
+lidar package
+==============
+
+Submodules
+----------
+
+lidar.filtering module
+----------------------
+
+.. automodule:: lidar.filtering
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+lidar.filling module
+--------------------
+
+.. automodule:: lidar.filling
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+lidar.slicing module
+--------------------
+
+.. automodule:: lidar.slicing
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+lidar.mounts module
+-------------------
+
+.. automodule:: lidar.mounts
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+
+
+Module contents
+---------------
+
+.. automodule:: lidar
+   :members:
+   :undoc-members:
+   :show-inheritance:
