### 2021-01-05 d1a0bd6
Added docstrings and reformatted code
