+++ lidar/toolbox/scripts/4_Slicing.py
@@ -1,4 +1,5 @@
 import sys
+
 # sys.path.append("./hill_shading")
 from skimage.external.tifffile import TiffFile
 import matplotlib.pyplot as plt
@@ -10,13 +11,26 @@ import math
 import time
 import os
 import shutil
-from osgeo import gdal,ogr,osr
+from osgeo import gdal, ogr, osr
+
 # from hillshade import hill_shade
 import arcpy
 
 ## class for true depression
 class Depression:
-    def __init__(self, id, level, size, volume, meanDepth, maxDepth, minElev, bndElev, inNbrId, nbrId=0):
+    def __init__(
+        self,
+        id,
+        level,
+        size,
+        volume,
+        meanDepth,
+        maxDepth,
+        minElev,
+        bndElev,
+        inNbrId,
+        nbrId=0,
+    ):
         self.id = id
         self.size = size
         self.level = level
@@ -32,8 +46,10 @@ class Depression:
 # get min and max elevation of a dem
 def get_min_max_nodata(img_array):
     max_elev = np.max(img_array)
-    nodata = pow(10, math.floor(math.log10(np.max(image)))+2) -1  # based on the max value of the image, assign no data value
-    image[image <= 0] = nodata  #change no data value
+    nodata = (
+        pow(10, math.floor(math.log10(np.max(image))) + 2) - 1
+    )  # based on the max value of the image, assign no data value
+    image[image <= 0] = nodata  # change no data value
     min_elev = np.min(img_array)
     return min_elev, max_elev, nodata
 
@@ -49,6 +65,7 @@ def regionGroup(img_array, min_size, no_data):
     label_objects, nb_labels = ndimage.label(image_cleaned)
     # nb_labels is the total number of objects. 0 represents background object.
     return label_objects, nb_labels
+
 
 # extract a subset of the raster
 def extractByRegion(img_array, obj_array, obj_id, no_data):
@@ -86,7 +103,7 @@ def writeRaster(arr, out_path, template):
     outdriver = gdal.GetDriverByName("GTiff")
     # http://www.gdal.org/gdal_8h.html
     # GDT_Byte = 1, GDT_UInt16 = 2, GDT_UInt32 = 4, GDT_Int32 = 5, GDT_Float32 = 6,
-    outdata   = outdriver.Create(str(out_path), rows, cols, 1, gdal.GDT_UInt32)
+    outdata = outdriver.Create(str(out_path), rows, cols, 1, gdal.GDT_UInt32)
     # Write the array to the file, which is the original array in this example
     outdata.GetRasterBand(1).WriteArray(arr)
     # Set a no data value if required
@@ -99,19 +116,21 @@ def writeRaster(arr, out_path, template):
 
 
 # raster to vector
-def polygonize(img,shp_path):
+def polygonize(img, shp_path):
     # mapping between gdal type and ogr field type
-    type_mapping = {gdal.GDT_Byte: ogr.OFTInteger,
-                    gdal.GDT_UInt16: ogr.OFTInteger,
-                    gdal.GDT_Int16: ogr.OFTInteger,
-                    gdal.GDT_UInt32: ogr.OFTInteger,
-                    gdal.GDT_Int32: ogr.OFTInteger,
-                    gdal.GDT_Float32: ogr.OFTReal,
-                    gdal.GDT_Float64: ogr.OFTReal,
-                    gdal.GDT_CInt16: ogr.OFTInteger,
-                    gdal.GDT_CInt32: ogr.OFTInteger,
[…]
