+++ tests/test_lidar.py
@@ -28,7 +28,7 @@ class TestLidar(unittest.TestCase):
         runner = CliRunner()
         result = runner.invoke(cli.main)
         assert result.exit_code == 0
-        assert 'lidar.cli.main' in result.output
-        help_result = runner.invoke(cli.main, ['--help'])
+        assert "lidar.cli.main" in result.output
+        help_result = runner.invoke(cli.main, ["--help"])
         assert help_result.exit_code == 0
-        assert '--help  Show this message and exit.' in help_result.output
+        assert "--help  Show this message and exit." in help_result.output
