### 2021-01-05 ccd6f2c
Added unit tests
