+++ tests/test_lidar.py
@@ -7,9 +7,16 @@
 import unittest
 from click.testing import CliRunner
 
-from lidar import lidar
+# from lidar import lidar
 from lidar import cli
 
+import os
+import pkg_resources
+import richdem as rd
+from scipy import ndimage
+import numpy as np
+import time
+import lidar
 
 class TestLidar(unittest.TestCase):
     """Tests for `lidar` package."""
@@ -23,12 +30,143 @@ class TestLidar(unittest.TestCase):
     def test_000_something(self):
         """Test something."""
 
-    def test_command_line_interface(self):
-        """Test the CLI."""
-        runner = CliRunner()
-        result = runner.invoke(cli.main)
-        assert result.exit_code == 0
-        assert "lidar.cli.main" in result.output
-        help_result = runner.invoke(cli.main, ["--help"])
-        assert help_result.exit_code == 0
-        assert "--help  Show this message and exit." in help_result.output
+    # def test_command_line_interface(self):
+    #     """Test the CLI."""
+    #     runner = CliRunner()
+    #     result = runner.invoke(cli.main)
+    #     assert result.exit_code == 0
+    #     assert "lidar.cli.main" in result.output
+    #     help_result = runner.invoke(cli.main, ["--help"])
+    #     assert help_result.exit_code == 0
+    #     assert "--help  Show this message and exit." in help_result.output
+
+    def test_mean_filter(self):
+
+        # identify the sample data directory of the package
+        package_name = "lidar"
+        data_dir = pkg_resources.resource_filename(package_name, "data/")
+        print("Sample data directory: {}".format(data_dir))
+
+        # use the sample dem. Change it to your own dem if needed
+        in_dem = os.path.join(data_dir, "dem.tif")
+        out_dir = os.path.join(os.path.expanduser("~"), "temp")
+        if not os.path.exists(out_dir):
+            os.mkdir(out_dir)
+
+        mean_dem = os.path.join(out_dir, "mean.tif")
+        mean = lidar.MeanFilter(in_dem, kernel_size=3)  
+        rd.SaveGDAL(mean_dem, mean)
+
+        self.assertTrue(os.path.exists(mean_dem))
+
+    def test_median_filter(self):
+
+        # identify the sample data directory of the package
+        package_name = "lidar"
+        data_dir = pkg_resources.resource_filename(package_name, "data/")
+        print("Sample data directory: {}".format(data_dir))
+
+        # use the sample dem. Change it to your own dem if needed
+        in_dem = os.path.join(data_dir, "dem.tif")
+        out_dir = os.path.join(os.path.expanduser("~"), "temp")
+        if not os.path.exists(out_dir):
+            os.mkdir(out_dir)
+
+        median_dem = os.path.join(out_dir, "median.tif")
+        median = lidar.MedianFilter(in_dem, kernel_size=3)  
+        rd.SaveGDAL(median_dem, median)
+
+        self.assertTrue(os.path.exists(median_dem))
+
+    def test_gaussian_filter(self):
+
+        # identify the sample data directory of the package
+        package_name = "lidar"
+        data_dir = pkg_resources.resource_filename(package_name, "data/")
+        print("Sample data directory: {}".format(data_dir))
+
+        # use the sample dem. Change it to your own dem if needed
+        in_dem = os.path.join(data_dir, "dem.tif")
+        out_dir = os.path.join(os.path.expanduser("~"), "temp")
+        if not os.path.exists(out_dir):
+            os.mkdir(out_dir)
+
+        gaussian_dem = os.path.join(out_dir, "gaussian.tif")
+        gaussian = lidar.GaussianFilter(in_dem,  sigma=1)  
[…]
