### 2021-01-05 #2896 @giswqs
@kbarnhart Thank you for handling the submission. I have revised the repo according to your comments. Thanks.

- Added docstrings to all functions and classes ([link](https://lidar.readthedocs.io/en/latest/lidar.html#submodules))
- Added tests ([link](https://github.com/giswqs/lidar/blob/master/tests/test_lidar.py)) and GitHub actions ([link](https://github.com/giswqs/lidar/runs/1651187012?check_suite_focus=true))
- Cited dependencies ([link](https://github.com/giswqs/lidar/blob/master/paper/paper.md#acknowledgments))

The [lidar/toolbox/scripts/1_Extract_Sink.py](https://github.com/giswqs/lidar/blob/master/lidar/toolbox/scripts/1_Extract_Sink.py) and [lidar/toolbox/scripts/4_Slicing.py](https://github.com/giswqs/lidar/blob/master/lidar/toolbox/scripts/4_Slicing.py) are intended as the arcpy equivalent of [lidar/filling.py](https://github.com/giswqs/lidar/blob/master/lidar/filling.py) and [lidar/slicing.py](https://github.com/giswqs/lidar/blob/master/lidar/slicing.py), respectively. The other five scripts (i.e., 2_Delineate_Catchment.py, 3_Flow_Path.py, 5_Catchment_Hierarchy.py, 6_Simulate Inundation.py, 7_Play_Animation.py) are for the arcpy toolbox only as they rely on some arcpy functions, which are not available outside arcpy.
