### 2021-01-05 #2896 @kbarnhart
@giswqs thanks for that update. Could you please recommend reviewers for this submission. A list of people who have volunteered to review for JOSS is listed above. Of course, don't feel constrained to that list.

In addition, is the Arc Spatial Analyst extension is the only Arc extension required to use the toolbox? I skimmed for `arcpy.CheckOutExtension` and only saw "Spatial". I want to make sure to find reviewers who have access to the correct Arc licenses.
