### 2021-01-05 #2896 @giswqs
@kbarnhart Here are some potential reviewers from that list with expertise in GIS and Python. And yes, the toolbox only requires the Spatial Analyst extension. Thanks.

- sgrieve
- Scivision
- juliohm
- sjsrey
- mikejohnson51
- nagellette
- Narayana-Rao
