### 2021-01-11 #2896 @laijingtao
@kbarnhart Hi Katy, I'm happy to review this submission. Looks like an interesting work. I don't have access to ArcGIS right now so I'll focus on the Python implementation. Hope that's okay.
