### 2021-01-12 #2896 @kbarnhart
@laijingtao thank you for your willingness to review and for letting me know about your Arc access. Focusing on the python implementation would be great and I'll make sure at least one other reviewer has access to Arc. I'll add you as a reviewer now, but won't make the JOSS review issue (which will have checkboxes to guide you through the review) until I've found another reviewer.
