### 2021-01-12 #2896 @kbarnhart
:wave: @sgrieve @mikejohnson51 @r-barnes would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

The submission is "lidar: A Python package for terrain and hydrological analysis using digital elevation models".

This submission includes an ArcGIS toolbox, as well as a python package that implements similar functionality. Thus, it is important that at least one (ideally all) reviewers have access to the proprietary environment ArcGIS and the Spatial Analyst extension. For reference, [here is a link](https://joss.readthedocs.io/en/latest/review_criteria.html#what-about-submissions-that-rely-upon-proprietary-languages-development-environments) to the JOSS documentation statement regarding submissions which include proprietary languages/environments.

This is a pre-review issue which is used to find reviewers. Once 2-3 reviewers are found, I'll start the review on a dedicated GitHub issue. At present we are asking reviewers to complete reviews in 6 weeks. If you are not able to review and have someone to recommend, please mention them here (when mentioning, please place a space after the @ of a github handle, for example, you would refer to me as "@ kbarnhart").

If you are interested in reviewing, I would recommend looking over the journal's [conflict of interest policy](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html?#joss-conflict-of-interest-policy) before the review process starts.

If you have any questions about the JOSS review process, please do not hesitate to reach out to me on this issue or at krbarnhart@usgs.gov
