### 2021-01-13 #2896 @sgrieve
@kbarnhart Afraid I don't have any spare capacity to review at the moment, and due to the pandemic I don't have access to an ArcGIS licence. London is supposed to come out of lockdown by mid Feb, so give me a shout then if you have not found anyone else and I might be able to take a look.
