### 2021-01-14 #2896 @cheginit
@kbarnhart I am a Python developer but don't have access to ArcGIS. So I can review the non-ArcGIS parts of the submission. Please, let me know if that works.
