### 2021-01-15 #2896 @r-barnes
I could take a look, but I don't have access to ArcGIS.
