### 2021-01-15 #2965 @whedon
[REVIEW]: lidar: A Python package for delineating nested surface depressions from digital elevation data
**Submitting author:** @giswqs (<a href="http://orcid.org/0000-0001-5437-4073">Qiusheng Wu</a>)
**Repository:** <a href="https://github.com/giswqs/lidar" target ="_blank">https://github.com/giswqs/lidar</a>
**Version:** 0.6.1

- [ ] ~~**Performance:** If there are any performance claims of the software, have they been confirmed?~~ (If there are no claims, please check off this item.)
