### 2021-01-22 #2965 @laijingtao
Hi Katy @kbarnhart, it seems that I can't click the checkbox and my names is not listed as one of the assignees. Also, I clicked the link for accepting invitation but it said that it could find the invitation. This is my first time reviewing for joss. Could you help me figure out what could be wrong?
