### 2021-02-03 #2965 @laijingtao
I've finished my review for this repository. This package meets all the review criteria and is suitable for publication. My only suggestion is the documentation of the core methods could be improved by adding detailed description and example inputs/outputs. Other parts of the documentation are all great.
