### 2021-02-04 #2965 @cheginit
I've been reviewing your submission and here are my comments so far:

1. The article reads well and states the package capabilities clearly. However, you wrote on line 6:

> lidar is a Python package for terrain and hydrological analysis using digital elevation models (DEMs).

Then on line 18 said:

> The lidar Python package implements the level-set method and makes it possible for delineating the nested hierarchy of surface depressions as well as elevated terrain features.

The first part makes the scope of applicability of the package very broad, terrain and hydrological analysis, and then second part says that the package mainly implements level-set method for delineating depressions. So I think you need to make the scope of applicability more clear in the summary part similar to the "Statement of Need" section where you correctly state that the package does some terrain analysis which can be used with other hydrology analysis tools.

2. Regarding the tests, I ran `coverage` and here is the result:

```bash
$ coverage run --source lidar -m unittest discover tests/
$ coverage report -m
Name                 Stmts   Miss  Cover   Missing
--------------------------------------------------
lidar/__init__.py        9      0   100%
lidar/cli.py             9      3    67%   11-13
lidar/example.py        29     29     0%   1-48
lidar/filling.py       151     12    92%   119, 274, 356-376
lidar/filtering.py      85     36    58%   54-56, 84-86, 114-116, 125-175
lidar/gui.py            62     50    19%   15-165, 169, 173, 177
lidar/lidar.py           1      1     0%   3
lidar/mounts.py         55     20    64%   54, 72, 102-133
lidar/slicing.py       375     49    87%   211-216, 262-287, 320, 363, 494-497, 598, 600, 705-722
--------------------------------------------------
TOTAL                  776    200    74%
```

Though not necessary, my suggestion is to add a coverage badge and try to increase the coverage percentage to around 90% at least.

3. I am not sure about the journal's policies regarding code quality, @kbarnhart, but, in my opinion, the code needs some polishing. Some files have `if-main` as well as commented out lines of code which shouldn't be in the published version of a code. If these kind of things matter, I can provide a more detailed list of things that needs to be addressed.
4. Regarding the documentation, I think the usage section is lacking. The `lidar` section is essentially a duplicate of all the other sections. The history section is empty as well.
