### 2021-02-05 #2965 @kbarnhart
@cheginit thanks for these thoughtful comments. @giswqs when you respond to them, please paste them into an in-repo issue and comment there. Include a link to this repo by including the text openjournals/joss-reviews/issues/2965. This keeps only the top-level discussion on this thread, and more detailed discussion on in-repo issue threads and links those issues with this one.

I broadly agree with @cheginit 's comments and will expand a bit on items 2 and 3. If further clarification from me is useful on any of these issues, just tag me here or in the in-repo issue.

2. My only addition regarding tests is to emphasize the distinction between tests of interface (e.g., if the right format input is provided, can it run without errors) and tests that verify the algorithms are behaving as expecting). My experience is that it is not too hard to create 100% code coverage that is just a test of the interface but much harder to reach a lower coverage level with algorithm verification. Both types of tests are important and having tests that verify the algorithms are implemented correctly are critical.

3. JOSS doesn't have any explicit statements about code quality. However, we do expect

    > In addition, JOSS requires that software should be feature-complete (i.e. no half-baked solutions) and designed for maintainable extension (not one-off modifications of existing tools). “Minor utility” packages, including “thin” API clients, and single-function packages are not acceptable.

    To that end, a user shouldn't be expected to modify the source code in order to use it to its full level of functionality. E.g., modifying the `if-main` block to use a different parameter value or input file. I don't think `if-main` blocks are inherently problematic, but if they exist, a user shouldn't need to edit the code in order to use them. And they should be documented.
