### 2021-02-05 giswqs/lidar#10 @giswqs
Addressing JOSS review comments
This issue is for addressing the JOSS reviewers' comments. https://github.com/openjournals/joss-reviews/issues/2965
