### 2021-02-05 #2965 @giswqs
@kbarnhart @cheginit @laijingtao Thank you all for your very helpful comments. I have opened an [issue](https://github.com/giswqs/lidar/issues/10) on the lidar repo. I will keep you posted on this thread once I resolve all the comments. Thanks.
