### 2021-02-10 b2ca611
Updated README and paper #10
