+++ .github/workflows/build.yaml
@@ -43,7 +43,7 @@ jobs:
                   pip install wheel
                   pip install --user --no-cache-dir Cython
                   pip install --user -r requirements.txt
-                  pip install pygdal==3.0.4.6
+                  pip install pygdal==3.2.1.6
               #   pip install --user -r requirements_dev.txt
               #   python -m pip install --upgrade pip
             - name: lidar-TEST
