+++ tests/test_lidar.py
@@ -30,16 +30,6 @@ class TestLidar(unittest.TestCase):
     def test_000_something(self):
         """Test something."""
 
-    # def test_command_line_interface(self):
-    #     """Test the CLI."""
-    #     runner = CliRunner()
-    #     result = runner.invoke(cli.main)
-    #     assert result.exit_code == 0
-    #     assert "lidar.cli.main" in result.output
-    #     help_result = runner.invoke(cli.main, ["--help"])
-    #     assert help_result.exit_code == 0
-    #     assert "--help  Show this message and exit." in help_result.output
-
     def test_mean_filter(self):
 
         # identify the sample data directory of the package
@@ -168,5 +158,4 @@ class TestLidar(unittest.TestCase):
         )
         self.assertTrue(os.path.exists(mount_id_path))
         self.assertTrue(os.path.exists(mount_level_path))
-    
 
