+++ lidar/filtering.py
@@ -93,7 +93,7 @@ def GaussianFilter(in_dem, sigma=1, out_file=None):
 
     Args:
         in_dem (str): File path to the input image.
-        kernel_size (int, optional): The size of the moving window. Defaults to 3.
+        sigma (int, optional): Standard deviation. Defaults to 1.
         out_file (str, optional): File path to the output image. Defaults to None.
 
     Returns:
@@ -116,60 +116,3 @@ def GaussianFilter(in_dem, sigma=1, out_file=None):
         return out_file
 
     return gau
-
-
-# #####################################  main script
-if __name__ == "__main__":
-
-    # identify the sample data directory of the package
-    package_name = "lidar"
-    data_dir = pkg_resources.resource_filename(package_name, "data/")
-    print("Sample data directory: {}".format(data_dir))
-
-    # use the sample dem. Change it to your own dem if needed
-    in_dem = os.path.join(data_dir, "dem.tif")
-    out_dir = os.path.join(os.path.expanduser("~"), "temp")
-
-    if not os.path.exists(out_dir):
-        os.mkdir(out_dir)
-
-    median_dem = os.path.join(out_dir, "median.tif")
-    gaussian_dem = os.path.join(out_dir, "gaussian.tif")
-    mean_dem = os.path.join(out_dir, "mean.tif")
-
-    dem = rd.LoadGDAL(in_dem)  # original dem
-
-    mean = MeanFilter(in_dem, kernel_size=3)
-    mean_diff = mean - dem
-    med = MedianFilter(in_dem, kernel_size=3)
-    med_diff = med - dem
-    gau = GaussianFilter(in_dem, sigma=1)
-    gau_diff = gau - dem
-
-    # plotting data
-    dem_fig = rd.rdShow(
-        dem, ignore_colours=[0], axes=False, cmap="jet", figsize=(6, 5.5)
-    )
-    mean_fig = rd.rdShow(
-        mean, ignore_colours=[0], axes=False, cmap="jet", figsize=(6, 5.5)
-    )
-    med_fig = rd.rdShow(
-        med, ignore_colours=[0], axes=False, cmap="jet", figsize=(6, 5.5)
-    )
-    gau_fig = rd.rdShow(
-        gau, ignore_colours=[0], axes=False, cmap="jet", figsize=(6, 5.5)
-    )
-    mean_diff_fig = rd.rdShow(
-        mean_diff, ignore_colours=[0], axes=False, cmap="jet", figsize=(6, 5.5)
-    )
-    med_diff_fig = rd.rdShow(
-        med_diff, ignore_colours=[0], axes=False, cmap="jet", figsize=(6, 5.5)
-    )
-    gau_diff_fig = rd.rdShow(
-        gau_diff, ignore_colours=[0], axes=False, cmap="jet", figsize=(6, 5.5)
-    )
-
-    # save results
-    rd.SaveGDAL(mean_dem, mean)
-    rd.SaveGDAL(median_dem, med)
-    rd.SaveGDAL(gaussian_dem, gau)
