+++ lidar/utilities.py
@@ -1,0 +1,295 @@
+import csv
+import os
+import shutil
+import subprocess
+import tarfile
+import urllib.request
+import zipfile
+
+
+def in_colab_shell():
+    """Tests if the code is being executed within Google Colab."""
+    try:
+        import google.colab  # pylint: disable=unused-variable
+
+        return True
+    except ImportError:
+        return False
+
+
+def is_drive_mounted():
+    """Checks whether Google Drive is mounted in Google Colab.
+
+    Returns:
+        bool: Returns True if Google Drive is mounted, False otherwise.
+    """
+    drive_path = "/content/drive/My Drive"
+    if os.path.exists(drive_path):
+        return True
+    else:
+        return False
+
+
+def download_from_url(url, out_file_name=None, out_dir=".", unzip=True, verbose=True):
+    """Download a file from a URL (e.g., https://github.com/giswqs/whitebox/raw/master/examples/testdata.zip)
+
+    Args:
+        url (str): The HTTP URL to download.
+        out_file_name (str, optional): The output file name to use. Defaults to None.
+        out_dir (str, optional): The output directory to use. Defaults to '.'.
+        unzip (bool, optional): Whether to unzip the downloaded file if it is a zip file. Defaults to True.
+        verbose (bool, optional): Whether to display or not the output of the function
+    """
+    in_file_name = os.path.basename(url)
+
+    if out_file_name is None:
+        out_file_name = in_file_name
+    out_file_path = os.path.join(os.path.abspath(out_dir), out_file_name)
+
+    if verbose:
+        print("Downloading {} ...".format(url))
+
+    try:
+        urllib.request.urlretrieve(url, out_file_path)
+    except Exception:
+        raise Exception("The URL is invalid. Please double check the URL.")
+
+    final_path = out_file_path
+
+    if unzip:
+        # if it is a zip file
+        if ".zip" in out_file_name:
+            if verbose:
+                print("Unzipping {} ...".format(out_file_name))
+            with zipfile.ZipFile(out_file_path, "r") as zip_ref:
+                zip_ref.extractall(out_dir)
+            final_path = os.path.join(
+                os.path.abspath(out_dir), out_file_name.replace(".zip", "")
+            )
+
+        # if it is a tar file
+        if ".tar" in out_file_name:
+            if verbose:
+                print("Unzipping {} ...".format(out_file_name))
+            with tarfile.open(out_file_path, "r") as tar_ref:
+                tar_ref.extractall(out_dir)
+            final_path = os.path.join(
+                os.path.abspath(out_dir), out_file_name.replace(".tart", "")
+            )
+
+    if verbose:
+        print("Data downloaded to: {}".format(final_path))
+
+    return
+
+
+def download_from_gdrive(gfile_url, file_name, out_dir=".", unzip=True, verbose=True):
+    """Download a file shared via Google Drive
+       (e.g., https://drive.google.com/file/d/18SUo_HcDGltuWYZs1s7PpOmOq_FvFn04/view?usp=sharing)
+
+    Args:
+        gfile_url (str): The Google Drive shared file URL
+        file_name (str): The output file name to use.
+        out_dir (str, optional): The output directory. Defaults to '.'.
+        unzip (bool, optional): Whether to unzip the output file if it is a zip file. Defaults to True.
+        verbose (bool, optional): Whether to display or not the output of the function
+    """
+    try:
+        from google_drive_downloader import GoogleDriveDownloader as gdd
+    except ImportError:
+        print("GoogleDriveDownloader package not installed. Installing ...")
+        subprocess.check_call(
[…]
