+++ lidar/__init__.py
@@ -3,12 +3,13 @@
 """Top-level package for lidar."""
 
 __author__ = """Qiusheng Wu"""
-__email__ = 'giswqs@gmail.com'
-__version__ = '0.5.1'
+__email__ = "giswqs@gmail.com"
+__version__ = "0.5.1"
 
 from .filling import ExtractSinks
 from .slicing import DelineateDepressions
 from .filtering import MeanFilter, MedianFilter, GaussianFilter
-from .mounts import  DelineateMounts
-from .gui import gui, GUI
+from .mounts import DelineateMounts
+from .gui import gui
+
 # from .mounts import DelineateMounts

[…]
