+++ .github/workflows/docs.yml
@@ -13,8 +13,17 @@ jobs:
                   python-version: 3.8
             - name: Install dependencies
               run: |
+                  sudo apt-add-repository ppa:ubuntugis/ubuntugis-unstable -y
+                  sudo apt-get -qq update
+                  sudo apt-get install gdal-bin libgdal-dev -y
+                  export CPLUS_INCLUDE_PATH=/usr/include/gdal
+                  export CPLUS_INCLUDE_PATH=/usr/include/gdal
+                  gdal-config --version
+                  gdalinfo --version
                   python -m pip install --upgrade pip
+                  pip install wheel
                   pip install --user --no-cache-dir Cython
                   pip install --user -r requirements.txt
+                  pip install pygdal==3.2.1.6
             - run: pip install mkdocs-material mkdocstrings mkdocs-git-revision-date-plugin mkdocs-jupyter mkdocs-pdf-export-plugin
             - run: mkdocs gh-deploy --force
