+++ docs/citations.md
@@ -1,0 +1,39 @@
+The level-set algorithm in the **lidar** package has been published in
+the following article:
+
+-   **Wu, Q.**, Lane, C.R., Wang, L., Vanderhoof, M.K., Christensen,
+    J.R., & Liu, H. (2019). Efficient Delineation of Nested Depression
+    Hierarchy in Digital Elevation Models for Hydrological Analysis
+    Using Level-Set Method. *Journal of the American Water Resources
+    Association*. DOI: [10.1111/1752-1688.12689](https://doi.org/10.1111/1752-1688.12689) ([PDF](https://spatial.utk.edu/pubs/2019_JAWRA.pdf))
+
+Applications of the level-set and contour-tree methods for feature
+extraction from LiDAR data:
+
+-   **Wu, Q.**, & Lane, C.R. (2017). Delineating wetland catchments and
+    modeling hydrologic connectivity using LiDAR data and aerial
+    imagery. *Hydrology and Earth System Sciences*. 21: 3579-3595. DOI:
+    [10.5194/hess-21-3579-2017](https://doi.org/10.5194/hess-21-3579-2017)
+-   **Wu, Q.**, Deng, C., & Chen, Z. (2016). Automated delineation of
+    karst sinkholes from LiDAR-derived digital elevation models.
+    *Geomorphology*. 266: 1-10. DOI:
+    [10.1016/j.geomorph.2016.05.006](http://dx.doi.org/10.1016/j.geomorph.2016.05.006)
+-   **Wu, Q.**, Su, H., Sherman, D.J., Liu, H., Wozencraft, J.M., Yu,
+    B., & Chen, Z. (2016). A graph-based approach for assessing
+    storm-induced coastal changes. *International Journal of Remote
+    Sensing*. 37:4854-4873. DOI:
+    [10.1080/01431161.2016.1225180](http://dx.doi.org/10.1080/01431161.2016.1225180)
+-   **Wu, Q.**, & Lane, C.R. (2016). Delineation and quantification of
+    wetland depressions in the Prairie Pothole Region of North Dakota.
+    *Wetlands*. 36(2):215–227. DOI:
+    [10.1007/s13157-015-0731-6](http://dx.doi.org/10.1007/s13157-015-0731-6)
+-   **Wu, Q.**, Liu, H., Wang, S., Yu, B., Beck, R., & Hinkel, K.
+    (2015). A localized contour tree method for deriving geometric and
+    topological properties of complex surface depressions based on
+    high-resolution topographic data. *International Journal of
+    Geographical Information Science*. 29(12): 2041-2060. DOI:
+    [10.1080/13658816.2015.1038719](http://dx.doi.org/10.1080/13658816.2015.1038719)
+-   **Wu, Q.**, Lane, C.R., & Liu, H. (2014). An effective method for
+    detecting potential woodland vernal pools using high-resolution
+    LiDAR data and aerial imagery. *Remote Sensing*. 6(11):11444-11467.
+    DOI: [10.3390/rs61111444](http://dx.doi.org/10.3390/rs61111444)
