+++ docs/usage.md
@@ -1,1 +1,17 @@
 # Usage
+
+The images below show working examples of the level set method for
+delineating nested depressions in the Cottonwood Lake Study Area (CLSA),
+North Dakota. More test datasets (e.g., the Pipestem watershed in the
+Prairie Pothole Region of North Dakota) can be downloaded from
+<http://gishub.org/2018-JAWRA-Data>
+
+The following example was conducted on a 64-bit Linux machine with a
+quad-core Intel i7-7700 CPU and 16 GB RAM. The average running time of
+the algorithm for this DEM was 0.75 seconds.
+
+![image](https://wetlands.io/file/images/CLSA_DEM.jpg)
+
+![image](https://wetlands.io/file/images/CLSA_DEM.jpg)
+
+![image](https://wetlands.io/file/images/CLSA_Table.jpg)
