### 2021-02-10 #2965 @giswqs
I have re-built the documentation using mkdocs. Here is the new link (https://lidar.gishub.org). I am still updating the content. The [lidar.reathedocs.io](https://lidar.readthedocs.io/) is no longer being updated. Thanks.
