+++ README.rst
@@ -1,370 +1,0 @@
-=====
-lidar
-=====
-
-.. image:: https://mybinder.org/badge_logo.svg 
-        :target: https://gishub.org/lidar-cloud
-
-.. image:: https://binder.pangeo.io/badge.svg 
-        :target: https://binder.pangeo.io/v2/gh/giswqs/lidar/master?filepath=examples%2Flidar.ipynb
-
-.. image:: https://img.shields.io/pypi/v/lidar.svg
-        :target: https://pypi.python.org/pypi/lidar
-
-.. image:: https://pepy.tech/badge/lidar
-        :target: https://pepy.tech/project/lidar
-
-.. image:: https://img.shields.io/conda/vn/conda-forge/lidar.svg
-        :target: https://anaconda.org/conda-forge/lidar
-
-.. image:: https://github.com/giswqs/lidar/workflows/build/badge.svg
-        :target: https://github.com/giswqs/lidar/actions?query=workflow%3Abuild
-
-.. image:: https://readthedocs.org/projects/lidar/badge/?version=latest
-        :target: https://lidar.readthedocs.io/en/latest/?badge=latest
-        :alt: Documentation Status
-
-.. image:: https://img.shields.io/badge/License-MIT-yellow.svg
-        :target: https://opensource.org/licenses/MIT
-
-.. image:: https://img.shields.io/twitter/follow/giswqs?style=social   
-        :target: https://twitter.com/giswqs
-
-.. image:: https://img.shields.io/badge/Donate-Buy%20me%20a%20coffee-yellowgreen.svg
-        :target: https://www.buymeacoffee.com/giswqs
-
-.. image:: https://joss.theoj.org/papers/005bf3e6f47840e74e71678d8e88facc/status.svg
-        :target: https://joss.theoj.org/papers/005bf3e6f47840e74e71678d8e88fac
-
-Author: Qiusheng Wu (https://wetlands.io)
-
-
-**lidar** is Python package for delineating the nested hierarchy of surface depressions 
-in digital elevation models (DEMs). It is particularly useful for analyzing high-resolution 
-topographic data, such as DEMs derived from Light Detection and Ranging (LiDAR) data.
-
-
-* GitHub repo: https://github.com/giswqs/lidar
-* Documentation: https://lidar.readthedocs.io
-* PyPI: https://pypi.org/project/lidar/
-* Binder: https://gishub.org/lidar-cloud
-* Free software: `MIT license`_
-
-.. _`MIT license`: https://en.wikipedia.org/wiki/MIT_License
-
-
-**Contents**
-
-- `Introduction`_
-- `Statement of Need`_
-- `Features`_
-- `Installation`_
-- `Tutorials`_
-- `lidar GUI`_
-- `Dependencies`_
-- `Examples`_
-- `References`_
-- `Reporting Bugs`_
-- `Credits`_
-
-
-Introduction
-------------
-
-**lidar** is a Python package for delineating the nested hierarchy of surface depressions 
-in digital elevation models (DEMs). In traditional hydrological modeling, surface depressions 
-in a DEM are commonly treated as artifacts and thus filled and removed to create a depressionless DEM,
-which can then be used to generate continuous stream networks. In reality, however, surface depressions 
-in DEMs are commonly a combination of spurious and actual terrain features. 
-Fine-resolution DEMs derived from Light Detection and Ranging (LiDAR) data can capture and represent 
-actual surface depressions, especially in glaciated and karst landscapes. 
-During the past decades, various algorithms have been developed to identify and delineate surface depressions, 
-such as depression filling, depression breaching, hybrid breaching-filling, 
-and contour tree method. More recently, a level-set method based on graph theory was proposed to delineate 
[…]
