+++ AUTHORS.rst
@@ -1,13 +1,0 @@
-=======
-Credits
-=======
-
-Development Lead
-----------------
-
-* Qiusheng Wu <giswqs@gmail.com>
-
-Contributors
-------------
-
-None yet. Why not be the first?
