+++ HISTORY.rst
@@ -1,20 +1,0 @@
-=======
-History
-=======
-0.5.1 (2020-12-04)
-------------------
-
-0.2.0 (2018-09-16)
-------------------
-
-0.1.6 (2018-05-21)
-------------------
-
-0.1.5 (2018-05-16)
-------------------
-
-0.1.3 (2018-05-15)
-------------------
-
-0.1.0 (2018-05-14)
-------------------

[…]
