+++ docs/index.md
@@ -18,7 +18,7 @@ particularly useful for analyzing high-resolution topographic data, such as DEMs
 -   Documentation: <https://lidar.gishub.org>
 -   PyPI: <https://pypi.org/project/lidar>
 -   Conda-forge: <https://anaconda.org/conda-forge/lidar>
--   Binder: <https://gishub.org/lidar-cloud>
+-   Open in Colab: <https://gishub.org/lidar-colab>
 -   Free software: [MIT license](https://opensource.org/licenses/MIT)
 
 ## Introduction
