+++ setup.py
@@ -86,6 +86,6 @@ setup(
     test_suite='tests',
     tests_require=test_requirements,
     url='https://github.com/giswqs/lidar',
-    version='0.5.1',
+    version='0.5.2',
     zip_safe=False,
 )
