+++ lidar/__init__.py
@@ -4,7 +4,7 @@
 
 __author__ = """Qiusheng Wu"""
 __email__ = "giswqs@gmail.com"
-__version__ = "0.5.1"
+__version__ = "__version__ = '0.5.2'"
 
 from .filling import ExtractSinks
 from .slicing import DelineateDepressions
