+++ setup.py
@@ -78,6 +78,7 @@ setup(
     dependency_links=dependency_links,
     license="MIT license",
     long_description=readme,
+    long_description_content_type='text/markdown',
     include_package_data=True,
     keywords='lidar',
     name='lidar',
