+++ docs/changelog.md
@@ -1,8 +1,11 @@
 # Changelog
 
+## v0.5.3 - February 10, 2021
+
+- Fixed PyPI markdown rendering error
 ## v0.5.2 - February 10, 2021
 
-- Added new documenation website (<https://lidar.gishub.org>)
+- Added new documentation website (<https://lidar.gishub.org>)
 - Improved JOSS paper
 - Cleaned up source code
 
