### 2021-02-11 #2965 @cheginit
@giswqs Thanks for taking the time and addressing the issues that I raised. @kbarnhart, in my opinion, the package meets all the criteria.

@giswqs As a side note, I would suggest you to take a look at [pre-commit](https://pre-commit.com/). It's a fantastic tool that can help with checking the quality of your code. It can automate many of the tasks, here's an example config file that I use for my package: https://github.com/cheginit/hydrodata/blob/master/.pre-commit-config.yaml
