### 2021-02-11 #2965 @giswqs
@cheginit Many thanks for your recommendation of [pre-commit](https://pre-commit.com/). It seems a very useful tool. I will start using it in my projects.
