+++ docs/get-started.md
@@ -52,6 +52,6 @@ lidar.gui()
 
 ## ArcGIS Toolbox
 
-[**Delineating nested surface depressions catchments - Part 1**](https://youtu.be/PpF8sfvCATE)
+[**Delineating nested surface depressions catchments using ArcMap**](https://youtu.be/PpF8sfvCATE)
 
 [![demo](http://img.youtube.com/vi/PpF8sfvCATE/0.jpg)](http://www.youtube.com/watch?v=PpF8sfvCATE)
