### 2021-02-27 giswqs/lidar#10 @amanaster2
**Installation instruction suggestions:**

- Need clarity regarding which method to use for installing (`mamba` vs. `pip`). I was not able to get the `mamba` installation of `lidar` working on either computer I tried it on. It would be safer to assume that folks don't have the necessary dependencies (i.e., GDAL, richdem) already installed, especially if they're creating a new environment.
- Include the _Dependencies_ section of documentation with the _Installation_ section for ease of use.
- Under _Dependencies_, the _Windows_ subsection is essentially a repeat of the initial command provided in the section. Suggest either removing one of the commands or making the commands the same. (Second command has `pip install richdem`, and I found that `pip install lidar` gives `richdem`, so that line isn't necessary.)
