### 2021-02-27 #2965 @amanaster2
Hello @kbarnhart, @giswqs:

I've finished my review. Overall, this is a great piece of software, and I really enjoyed learning how to use it.
I have a few comments/suggestions:

1.  **Documentation**: The installation instructions could use clarification. Please see my comment on this [issue](https://github.com/giswqs/lidar/issues/10) for more details.
2. **Documentation**: Suggest providing information about the ArcGIS toolbox in the [lidar repository](https://github.com/giswqs/lidar). I wouldn't have known there was an ArcGIS toolbox at first glance had @kbarnhart not told me. The tutorials you provided were excellent, and the toolbox works great!
3.  **Paper**: A little more information regarding the state of the field would be ideal. Are there other software packages that are comparable?
4. **Paper**:  Update the paper such that the readthedocs isn't listed under Help Documentation and instead list the [gishub](https://lidar.gishub.org/) page.

Please let me know if you have any questions about any of my comments.
Otherwise, this software meets all review criteria, and I would be happy to see it published!
