+++ docs/get-started.md
@@ -52,6 +52,14 @@ lidar.gui()
 
 ## ArcGIS Toolbox
 
+### Toolbox interface
+
+![toolbox](https://raw.githubusercontent.com/giswqs/lidar/master/images/toolbox_0.png)
+
+![toolbox_ui](https://raw.githubusercontent.com/giswqs/lidar/master/images/toolbox_ui.png)
+
+### Video tutorials
+
 [**Delineating nested surface depressions and catchments using ArcGIS Pro**](https://youtu.be/PpF8sfvCATE)
 
 [![demo](http://img.youtube.com/vi/W9PFHNV3cT0/0.jpg)](http://www.youtube.com/watch?v=W9PFHNV3cT0)
