+++ paper/paper.bib
@@ -239,3 +239,13 @@
   issn        = "2167-9843",
   doi         = "10.7287/peerj.preprints.27099v1"
 }
+
+@MISC{Lindsay2018,
+  title        = "{WhiteboxTools User Manual}",
+  author       = "Lindsay, John B",
+  publisher    = "GitHub.com",
+  year         =  2018,
+  url          = "https://jblindsay.github.io/wbt_book",
+  howpublished = "\url{https://jblindsay.github.io/wbt_book}",
+  note         = "Accessed: 2021-1-7"
+}
