+++ paper/paper.bib
@@ -246,6 +246,5 @@
   publisher    = "GitHub.com",
   year         =  2018,
   url          = "https://jblindsay.github.io/wbt_book",
-  howpublished = "\url{https://jblindsay.github.io/wbt_book}",
   note         = "Accessed: 2021-1-7"
 }
