### 2021-02-27 #2965 @giswqs
@amanaster2 Thank you very much for taking the time to review the software and providing feedback. I have updated the repo according to your comments. Thanks.

1. Improved the installation instructions. See https://lidar.gishub.org/installation
2. Added the ArcGIS toolbox info. See https://lidar.gishub.org/get-started/#arcgis-toolbox
3. Added a `State of the field` section to the paper. See the [article proof.](https://github.com/openjournals/joss-papers/blob/joss.02965/joss.02965/10.21105.joss.02965.pdf)
4. Updated the `Help Documentation` of the paper. See the [article proof.](https://github.com/openjournals/joss-papers/blob/joss.02965/joss.02965/10.21105.joss.02965.pdf)

@kbarnhart Please note that I changed the paper title to **lidar: A Python package for delineating nested surface depressions from digital elevation data** based on cheginit's comments.
