### 2021-03-01 #2965 @kbarnhart
@giswqs thanks for making changes in response to all of the reviewer's comments.

@laijingtao @cheginit @amanaster2 - if @giswqs has addressed all of your concerns and you think that this submission is ready to be accepted, could you please indicate so here. If not, please let me know what else needs to be addressed.

@giswqs , after all reviewer's have recommended the submission be accepted, there are a couple of steps I complete (e.g., copyediting) before we move to the final steps of the JOSS process.
