### 2021-03-01 #2965 @kbarnhart
@giswqs please fix the .bib file so John B Lindsay and J B Lindsay are not disambiguated. This occurs when the author names are different. If you make all of them "Lindsay, J B" it should fix it.

Once you have made that fix and verified that it builds correctly as a PDF, please do the following:
