+++ paper/paper.bib
@@ -242,7 +242,7 @@
 
 @MISC{Lindsay2018,
   title        = "{WhiteboxTools User Manual}",
-  author       = "Lindsay, John B",
+  author       = "Lindsay, J B",
   publisher    = "GitHub.com",
   year         =  2018,
   url          = "https://jblindsay.github.io/wbt_book",
