+++ paper/paper.bib
@@ -104,6 +104,7 @@
   title     = "{Modelling surface drainage patterns in altered landscapes using
                {LiDAR}}",
   author    = "Lindsay, J B and Dhun, K",
+  journal   = "International Journal of Geographical Information Science",
   publisher = "Taylor \& Francis",
   volume    =  29,
   number    =  3,
@@ -226,6 +227,7 @@
 @ARTICLE{Barnes2018,
   title       = "{RichDEM: High-performance terrain analysis}",
   author      = "Barnes, Richard",
+  journal     = "PeerJ Preprints",
   publisher   = "PeerJ Inc.",
   number      = "e27099v1",
   institution = "PeerJ Preprints",
