+++ setup.py
@@ -87,6 +87,6 @@ setup(
     test_suite='tests',
     tests_require=test_requirements,
     url='https://github.com/giswqs/lidar',
-    version='0.6.0',
+    version='0.6.1',
     zip_safe=False,
 )
