+++ setup.cfg
@@ -1,5 +1,5 @@
 [bumpversion]
-current_version = 0.6.0
+current_version = 0.6.1
 commit = True
 tag = True
 
@@ -18,4 +18,3 @@ universal = 1
 exclude = docs
 
 [aliases]
-
