### 2021-03-01 #2965 @giswqs
@kbarnhart I have fixed the paper.bib file. See the [article proof](https://github.com/openjournals/joss-papers/blob/joss.02965/joss.02965/10.21105.joss.02965.pdf). Thanks.

- Tagged version: [v0.6.1](https://github.com/giswqs/lidar/releases/tag/v0.6.1)
- Zenodo link: https://doi.org/10.5281/zenodo.4571011
- DOI: 10.5281/zenodo.4571011
