### 2021-03-02 #2965 @kbarnhart
@giswqs I've now recommended that this submission be accepted and published. One of the JOSS EiC's @openjournals/joss-eics will manage final publication steps.

Thanks to @laijingtao @cheginit and @amanaster2 for contributing reviews.
